pyproject.toml
[project]
name = "ai-pre-sales-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.59",
    "chromadb>=1.5.2",
    "fastapi>=0.134.0",
    "fitz>=0.0.1.dev2",
    "llama-index>=0.12.52",
    "llama-index-embeddings-bedrock>=0.4.0",
    "llama-index-llms-bedrock>=0.3.8",
    "llama-index-vector-stores-chroma>=0.4.2",
    "openpyxl>=3.1.5",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
